[Update] Move utils function into index.js
