--- src/utils/get-embedded-query-param.js
@@ -1,10 +1,0 @@
-'use strict'
-
-export function getArticleEmbeddedQuery() {
-  const list = [ 'authorImages', 'og_image', 'tags', 'categories' ]
-  let query = {}
-  list.forEach((ele) => {
-    query[ele] = 1
-  })
-  return query
-}
