--- src/utils/get-list-id.js
@@ -1,14 +1,0 @@
-/*global __DEVELOPMENT__ */
-'use strict'
-import { devCatListId, prodCatListId } from '../conf/list-id'
-
-function getCatId(catName) {
-  if (__DEVELOPMENT__) {
-    return devCatListId[catName]
-  }
-  return prodCatListId[catName]
-}
-
-export {
-  getCatId
-}
